+++ test/io_tests/prog_test.py
@@ -1,30 +1,13 @@
 """Unit test for KNX/IP gateway scanner."""
 import asyncio
-from unittest.mock import MagicMock, create_autospec, patch
+from unittest.mock import patch
 
 import pytest
 from xknx import XKNX
-from xknx.io import GatewayScanFilter, GatewayScanner, UDPClient
-from xknx.io.gateway_scanner import GatewayDescriptor
-from xknx.knxip import (
-    HPAI,
-    DIBDeviceInformation,
-    DIBServiceFamily,
-    DIBSuppSVCFamilies,
-    KNXIPFrame,
-    SearchResponse,
-)
-from xknx.telegram import (
-    IndividualAddress,
-    Priority,
-    Telegram,
-    TelegramDirection,
-    TPDUType,
-)
-from xknx.telegram.apci import (
-    IndividualAddressResponse,
-)
-from xknx.prog.management import NM_EXISTS, NM_OK, NM_TIME_OUT, NetworkManagement
+#from xknx.io import GatewayScanFilter, GatewayScanner, UDPClient
+#from xknx.io.gateway_scanner import GatewayDescriptor
+from xknx.telegram import IndividualAddress
+from xknx.prog.management import NM_EXISTS, NM_OK, NetworkManagement
 
 
 @pytest.mark.asyncio
@@ -32,30 +15,31 @@ class TestProgrammingInterface:
     """Test class for xknx/io/GatewayScanner objects."""
 
     called = 0
-    async def fake_DeviceDescriptor_Read_Response(self, *p):
+    async def fake_devicedescriptor_read_response(self, *p):
+        """Fake function for evicedescriptor_read_response."""
         if self.called == 1:
             return
         self.called+=1
         await asyncio.sleep(60.)
 
-    @patch("xknx.prog.device.A_Device.IndividualAddress_Read_Response", autospec=True)
-    @patch("xknx.prog.device.A_Device.DeviceDescriptor_Read_Response", autospec=True)
-    async def test_write_individual_address_success(self, mock_DeviceDescriptor_Read_Response, mock_IndividualAddress_Read_Response):
-        """Test finding all valid interfaces to send search requests to. No requests are sent."""
+    @patch("xknx.prog.device.ProgDevice.individualaddress_read_response", autospec=True)
+    @patch("xknx.prog.device.ProgDevice.devicedescriptor_read_response", autospec=True)
+    async def test_write_individual_address_success(self, mock_devicedescriptor_read_response, mock_individualaddress_read_response):
+        """Test IndividualAddress_Write with success."""
         xknx = XKNX()
-        mock_DeviceDescriptor_Read_Response.side_effect = self.fake_DeviceDescriptor_Read_Response
+        mock_devicedescriptor_read_response.side_effect = self.fake_devicedescriptor_read_response
         network_management = NetworkManagement(xknx)
-        return_code = await network_management.IndividualAddress_Write(
+        return_code = await network_management.individualaddress_write(
             IndividualAddress("1.2.1")
             )
         assert return_code == NM_OK
 
-    @patch("xknx.prog.device.A_Device.DeviceDescriptor_Read_Response", autospec=True)
-    async def test_write_individual_address_exists(self, mock_DeviceDescriptor_Read_Response):
-        """Test finding all valid interfaces to send search requests to. No requests are sent."""
+    @patch("xknx.prog.device.ProgDevice.devicedescriptor_read_response", autospec=True)
+    async def test_write_individual_address_exists(self, mock_devicedescriptor_read_response):
+        """Test IndividualAddress_Write with existing device."""
         xknx = XKNX()
         network_management = NetworkManagement(xknx)
-        return_code = await network_management.IndividualAddress_Write(
+        return_code = await network_management.individualaddress_write(
             IndividualAddress("1.2.1")
             )
         assert return_code == NM_EXISTS