+++ test/str_test.py
@@ -494,7 +494,8 @@ class TestStringRepresentations:
     def test_hpai(self):
         """Test string representation of HPAI."""
         hpai = HPAI(ip_addr="192.168.42.1", port=33941)
-        assert str(hpai) == "<HPAI 192.168.42.1:33941 />"
+        assert str(hpai) == "192.168.42.1:33941"
+        assert repr(hpai) == "HPAI('192.168.42.1', 33941)"
 
     def test_header(self):
         """Test string representation of KNX/IP-Header."""
@@ -515,7 +516,7 @@ class TestStringRepresentations:
         connect_request.data_endpoint = HPAI(ip_addr="192.168.42.2", port=33942)
         assert (
             str(connect_request)
-            == '<ConnectRequest control_endpoint="<HPAI 192.168.42.1:33941 />" data_endpoint="<HPAI 192.168.42.2:33942 />" '
+            == '<ConnectRequest control_endpoint="192.168.42.1:33941" data_endpoint="192.168.42.2:33942" '
             'request_type="ConnectRequestType.TUNNEL_CONNECTION" />'
         )
 
@@ -530,7 +531,7 @@ class TestStringRepresentations:
         assert (
             str(connect_response)
             == '<ConnectResponse communication_channel="13" status_code="ErrorCode.E_NO_ERROR" '
-            'control_endpoint="<HPAI 192.168.42.1:33941 />" '
+            'control_endpoint="192.168.42.1:33941" '
             'request_type="ConnectRequestType.TUNNEL_CONNECTION" identifier="42" />'
         )
 
@@ -542,7 +543,7 @@ class TestStringRepresentations:
         disconnect_request.control_endpoint = HPAI(ip_addr="192.168.42.1", port=33941)
         assert (
             str(disconnect_request)
-            == '<DisconnectRequest CommunicationChannelID="13" control_endpoint="<HPAI 192.168.42.1:33941 />" />'
+            == '<DisconnectRequest CommunicationChannelID="13" control_endpoint="192.168.42.1:33941" />'
         )
 
     def test_disconnect_response(self):
@@ -565,7 +566,7 @@ class TestStringRepresentations:
         )
         assert (
             str(connectionstate_request)
-            == '<ConnectionStateRequest CommunicationChannelID="23", control_endpoint="<HPAI 192.168.42.1:33941 />" />'
+            == '<ConnectionStateRequest CommunicationChannelID="23", control_endpoint="192.168.42.1:33941" />'
         )
 
     def test_connectionstate_response(self):
@@ -584,7 +585,7 @@ class TestStringRepresentations:
         search_request = SearchRequest(xknx)
         assert (
             str(search_request)
-            == '<SearchRequest discovery_endpoint="<HPAI 224.0.23.12:3671 />" />'
+            == '<SearchRequest discovery_endpoint="224.0.23.12:3671" />'
         )
 
     def test_search_response(self):
@@ -596,7 +597,7 @@ class TestStringRepresentations:
         search_response.dibs.append(DIBGeneric())
         assert (
             str(search_response)
-            == '<SearchResponse control_endpoint="<HPAI 192.168.42.1:33941 />" dibs="[\n'
+            == '<SearchResponse control_endpoint="192.168.42.1:33941" dibs="[\n'
             '<DIB dtc="0" data="" />,\n'
             '<DIB dtc="0" data="" />\n'
             ']" />'
@@ -650,7 +651,7 @@ class TestStringRepresentations:
             str(knxipframe)
             == '<KNXIPFrame <KNXIPHeader HeaderLength="6" ProtocolVersion="16" KNXIPServiceType="SEARCH_REQUEST" '
             'Reserve="0" TotalLength="0" />\n'
-            ' body="<SearchRequest discovery_endpoint="<HPAI 224.0.23.12:3671 />" />" />'
+            ' body="<SearchRequest discovery_endpoint="224.0.23.12:3671" />" />'
         )
 
     #