+++ xknx/prog/management.py
@@ -5,17 +5,15 @@ in KNX-Standard 3.5.2
 
 import asyncio
 
-from _testcapi import awaitType
-from xknx import telegram
-from xknx.telegram.address import GroupAddress, IndividualAddress
-
-from xknx.prog.device import A_Device
+from xknx.telegram.address import GroupAddress
+from xknx.prog.device import ProgDevice
 
 NM_OK = 0
 NM_EXISTS = 1
 NM_TIME_OUT = 2
 
 class NetworkManagement:
+    """Class for network management functionality."""
 
     def __init__(self, xknx):
         self.xknx = xknx
@@ -25,50 +23,50 @@ class NetworkManagement:
         # map for registered devices
         self.reg_dev = {}
 
-    async def telegram_received_cb(self, telegram):
+    async def telegram_received_cb(self, tele):
         """Do something with the received telegram."""
-        print(f"Telegram received: {telegram}")
-        if telegram.source_address in self.reg_dev:
-            await self.reg_dev[telegram.source_address].process_telegram(telegram)
-        if telegram.destination_address == GroupAddress("0"):
-            for addr in self.reg_dev:
-                await self.reg_dev[addr].process_telegram(telegram)
+        if tele.source_address in self.reg_dev:
+            await self.reg_dev[tele.source_address].process_telegram(tele)
+        if tele.destination_address == GroupAddress("0"):
+            for reg_dev_val in self.reg_dev.values():
+                await reg_dev_val.process_telegram(tele)
 
-    async def IndividualAddress_Write(self, ia):
-        d = A_Device(self.xknx, ia)
-        self.reg_dev[ia] = d
+    async def individualaddress_write(self, ind_add):
+        """Perform IndividualAdress_Write."""
+        device = ProgDevice(self.xknx, ind_add)
+        self.reg_dev[ind_add] = device
 
         # chech if IA is already present
         try:
-            await asyncio.wait_for(d.T_Connect_Response(), 0.5)
+            await asyncio.wait_for(device.t_connect_response(), 0.5)
             return NM_EXISTS
         except asyncio.TimeoutError:
             pass
 
         try:
-            await asyncio.wait_for(d.DeviceDescriptor_Read_Response(0), 0.5)
-            await d.T_Disconnect()
+            await asyncio.wait_for(device.devicedescriptor_read_response(0), 0.5)
+            await device.t_disconnect()
             return NM_EXISTS
         except asyncio.TimeoutError:
             pass
 
         # wait until PROG button is pressed
         try:
-            await asyncio.wait_for(d.IndividualAddress_Read_Response(), 600)
+            await asyncio.wait_for(device.individualaddress_read_response(), 600)
         except asyncio.TimeoutError:
             return NM_TIME_OUT
 
-        await d.IndividualAddress_Write()
+        await device.individualaddress_write()
 
         # Zusatz ETS reengeneering
-        await d.T_Connect()
+        await device.t_connect()
         try:
-            await asyncio.wait_for(d.DeviceDescriptor_Read_Response(0), 1.0)
+            await asyncio.wait_for(device.devicedescriptor_read_response(0), 1.0)
         except asyncio.TimeoutError:
-            raise RuntimeError(f"No device response from {ia}")
+            raise RuntimeError(f"No device response from {ind_add}")
 
-        await d.PropertyValue_Read()
-        await d.Restart()
+        await device.propertyvalue_read()
+        await device.restart()
         await asyncio.sleep(1)
-        await d.T_Disconnect()
+        await device.t_disconnect()
         return NM_OK